basic functionality for user stories completed

+++ AppLibrary/Seat.cs
@@ -8,10 +8,10 @@ namespace AppLibrary
     {
         public int Id { get; set; }
         public string UniqueId { get; set; }
-        public decimal Cost { get; set; }
+        public double Cost { get; set; }
         [Column(TypeName = "varchar(5)")]
         public string Occupant { get; set; } //type of occupant, "none", "adult", or "child"      
         public Section Section { get; set; }
-  
+
     }
 }
